small code layout changes

+++ sendpoint/main.c
@@ -13,18 +13,18 @@
 
 void printInvalidCoord(char coord);
 void printInstruction();
-int parseCoordinate(char * str, long *val);
-void sendPointData(char* url, char* data);
+int parseCoordinate(char *str, long *val);
+void sendPointData(char *url, char *data);
 
 int main(int argc, char * argv[]) {
     
     if(argc == 4){
-
+        
         char *jsonObj = calloc(1024, sizeof(char));
         char *url = calloc(1024, sizeof(char));
         
-        long *x = calloc(1, sizeof(int));
-        long *y = calloc(1, sizeof(int));
+        long *x = calloc(1, sizeof(long));
+        long *y = calloc(1, sizeof(long));
         
         if(parseCoordinate(argv[1], x)){
             printInvalidCoord('X');
@@ -37,7 +37,7 @@ int main(int argc, char * argv[]) {
         
         url = argv[3];
         
-        sprintf(jsonObj, "{\"X\":%lu, \"Y\":%lu}", *x, *y);
+        sprintf(jsonObj, "{\"X\":%ld, \"Y\":%ld}", *x, *y);
         
         sendPointData(url, jsonObj);
         
@@ -51,8 +51,8 @@ int main(int argc, char * argv[]) {
     }
     
     
-     return EXIT_SUCCESS;
-
+    return EXIT_SUCCESS;
+    
 }
 
 void printInvalidCoord(char coord){
@@ -65,34 +65,34 @@ void printInstruction(){
 }
 
 
-int parseCoordinate(char * str, long *val){
+int parseCoordinate(char *str, long *val){
     
-        int base = 10, returnValue = 0;
-        char *endptr;
-        
-        errno = 0;
-        *val = strtol(str, &endptr, base);
-        
-        if ((errno == ERANGE && (*val == LONG_MAX || *val == LONG_MIN))
-            || (errno != 0 && *val == 0)) {
-            returnValue = 1;
-        }
-        
-        if (endptr == str) {
-            returnValue = 1;
-        }
-        
-        if (*endptr != '\0'){
-            returnValue = 1;
-        }
-        
+    int base = 10, returnValue = 0;
+    char *endptr;
+    
+    errno = 0;
+    *val = strtol(str, &endptr, base);
+    
+    if ((errno == ERANGE && (*val == LONG_MAX || *val == LONG_MIN))
+        || (errno != 0 && *val == 0)) {
+        returnValue = 1;
+    }
+    
+    if (endptr == str) {
+        returnValue = 1;
+    }
+    
+    if (*endptr != '\0'){
+        returnValue = 1;
+    }
+    
     return returnValue;
 }
 
 
 
 
-void sendPointData(char* url, char* data){
+void sendPointData(char *url, char* data){
     CURL *curl;
     CURLcode res;
     
@@ -109,7 +109,7 @@ void sendPointData(char* url, char* data){
         curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers);
         curl_easy_setopt(curl, CURLOPT_URL, url);
         curl_easy_setopt(curl, CURLOPT_POSTFIELDS, data);
-
+        
         res = curl_easy_perform(curl);
         if(res != CURLE_OK){
             fprintf(stderr, "%s\n", curl_easy_strerror(res));
